+++ common/promapi.go
@@ -2,11 +2,12 @@ package common
 
 import "github.com/prometheus/prometheus/pkg/textparse"
 
-type APIResponse struct {
+type MetadataAPIResponse struct {
 	Status    string        `json:"status"`
-	Data      []APIMetadata `json:"data"`
-	Error     string        `json:"error"`
-	ErrorType string        `json:"errorType"`
+	Data      []APIMetadata `json:"data,omitempty"`
+	Error     string        `json:"error,omitempty"`
+	ErrorType string        `json:"errorType,omitempty"`
+	Warnings  []string      `json:"warnings,omitempty"`
 }
 
 type APIMetadata struct {
@@ -14,4 +15,16 @@ type APIMetadata struct {
 	Metric string               `json:"metric"`
 	Help   string               `json:"help"`
 	Type   textparse.MetricType `json:"type"`
+}
+
+type ConfigAPIResponse struct {
+	Status    string    `json:"status"`
+	Data      APIConfig `json:"data,omitempty"`
+	ErrorType string    `json:"errorType,omitempty"`
+	Error     string    `json:"error,omitempty"`
+	Warnings  []string  `json:"warnings,omitempty"`
+}
+
+type APIConfig struct {
+	YAML string `json:"yaml"`
 }