+++ metadata/metadata_test.go
@@ -30,6 +30,8 @@ import (
 	"github.com/prometheus/prometheus/pkg/textparse"
 )
 
+type metadataResponse common.MetadataAPIResponse
+
 func TestCache_Get(t *testing.T) {
 	metrics := []common.APIMetadata{
 		{Metric: "metric1", Type: textparse.MetricTypeCounter, Help: "help_metric1"},
@@ -40,7 +42,7 @@ func TestCache_Get(t *testing.T) {
 		{Metric: "metric6", Type: config.MetricTypeUntyped, Help: "help_metric6"},
 		{Metric: "metric_with_override", Type: textparse.MetricTypeGauge, Help: "help_metric_with_override"},
 	}
-	var handler func(qMetric, qMatch string) *common.APIResponse
+	var handler func(qMetric, qMatch string) *metadataResponse
 
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		err := json.NewEncoder(w).Encode(handler(
@@ -76,14 +78,14 @@ func TestCache_Get(t *testing.T) {
 	c := NewCache(nil, u, staticMetadata)
 
 	// First get for the job, we expect an initial batch request.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		if qMetric != "" {
 			t.Fatalf("unexpected metric %v in request", qMetric)
 		}
 		if qMatch != `{job="prometheus",instance="localhost:9090"}` {
 			t.Fatalf("unexpected matcher %v in request", qMatch)
 		}
-		return &common.APIResponse{Status: "success", Data: metrics[:4]}
+		return &metadataResponse{Status: "success", Data: metrics[:4]}
 	}
 	md, err := c.Get(ctx, "prometheus", "localhost:9090", "metric2")
 	if err != nil {
@@ -92,7 +94,7 @@ func TestCache_Get(t *testing.T) {
 	expect(metrics[1], md)
 
 	// Query metric that should have been retrieved in the initial batch.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		t.Fatal("unexpected request")
 		return nil
 	}
@@ -109,14 +111,14 @@ func TestCache_Get(t *testing.T) {
 	expect(metrics[2], md)
 
 	// Query metric that was not in the batch, expect a single-metric query.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		if qMetric != "metric5" {
 			t.Fatalf("unexpected metric %v in request", qMetric)
 		}
 		if qMatch != `{job="prometheus",instance="localhost:9090"}` {
 			t.Fatalf("unexpected matcher %v in request", qMatch)
 		}
-		return &common.APIResponse{Status: "success", Data: metrics[4:5]}
+		return &metadataResponse{Status: "success", Data: metrics[4:5]}
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "metric5")
 	if err != nil {
@@ -124,7 +126,7 @@ func TestCache_Get(t *testing.T) {
 	}
 	expect(metrics[4], md)
 	// It should be in our cache afterwards.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		t.Fatal("unexpected request")
 		return nil
 	}
@@ -135,14 +137,14 @@ func TestCache_Get(t *testing.T) {
 	expect(metrics[4], md)
 
 	// Test "untyped" metric type from Prometheus 2.4.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		if qMetric != "metric6" {
 			t.Fatalf("unexpected metric %v in request", qMetric)
 		}
 		if qMatch != `{job="prometheus",instance="localhost:9090"}` {
 			t.Fatalf("unexpected matcher %v in request", qMatch)
 		}
-		return &common.APIResponse{Status: "success", Data: metrics[5:6]}
+		return &metadataResponse{Status: "success", Data: metrics[5:6]}
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "metric6")
 	if err != nil {
@@ -163,14 +165,14 @@ func TestCache_Get(t *testing.T) {
 	}
 
 	// If a metric does not exist, we first expect a fetch attempt.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		if qMetric != "does_not_exist" {
 			t.Fatalf("unexpected metric %v in request", qMetric)
 		}
 		if qMatch != `{job="prometheus",instance="localhost:9090"}` {
 			t.Fatalf("unexpected matcher %v in request", qMatch)
 		}
-		return &common.APIResponse{Status: "error", ErrorType: apiErrorNotFound, Error: "does not exist"}
+		return &metadataResponse{Status: "error", ErrorType: apiErrorNotFound, Error: "does not exist"}
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "does_not_exist")
 	if err != nil {
@@ -180,7 +182,7 @@ func TestCache_Get(t *testing.T) {
 		t.Fatalf("expected nil metadata but got %v", md)
 	}
 	// Requesting it again should not do another request (modulo timeout).
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		t.Fatal("unexpected request")
 		return nil
 	}
@@ -193,7 +195,7 @@ func TestCache_Get(t *testing.T) {
 	}
 
 	// Test matcher escaping.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		if qMatch != `{job="prometheus\nwith_newline",instance="localhost:9090"}` {
 			t.Fatalf("matcher not escaped properly: %s", qMatch)
 		}
@@ -205,7 +207,7 @@ func TestCache_Get(t *testing.T) {
 	}
 
 	// Test fallthrough to static metadata.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		return nil
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "static_metric2")
@@ -218,8 +220,8 @@ func TestCache_Get(t *testing.T) {
 	}
 
 	// Test override with static metadata.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
-		return &common.APIResponse{Status: "success", Data: metrics}
+	handler = func(qMetric, qMatch string) *metadataResponse {
+		return &metadataResponse{Status: "success", Data: metrics}
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "metric_with_override")
 	if err != nil {
@@ -231,7 +233,7 @@ func TestCache_Get(t *testing.T) {
 	}
 
 	// Test recording rule.
-	handler = func(qMetric, qMatch string) *common.APIResponse {
+	handler = func(qMetric, qMatch string) *metadataResponse {
 		return nil
 	}
 	md, err = c.Get(ctx, "prometheus", "localhost:9090", "some:recording:rule")